requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
pyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
ensure==1.0.2
python-box==6.0.2
fastapi
uvicorn
Jinja2

-e .